stage1/init: some very simple code cleanup
